# Todos.ConsoleClient/Connected Services/TodoService/OpenAPI.nswag.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Todos Web API v1",
    "description": "Todos API",
    "contact": {
      "name": "ACME Company",
      "url": "https://todos.acme.com/",
      "email": ""
    },
    "version": "v1"
  },
  "paths": {
    "/api/Todo": {
      "get": {
        "tags": [
          "Todo"
        ],
        "operationId": "TodoAll",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TodoEntity"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TodoEntity"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TodoEntity"
                  }
                }
              },
              "application/x-yaml": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TodoEntity"
                  }
                }
              },
              "text/yaml": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TodoEntity"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Todo"
        ],
        "operationId": "Todo",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TodoEntity"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/TodoEntity"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/TodoEntity"
              }
            },
            "application/x-yaml": {
              "schema": {
                "$ref": "#/components/schemas/TodoEntity"
              }
            },
            "text/yaml": {
              "schema": {
                "$ref": "#/components/schemas/TodoEntity"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/TodoEntity"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TodoEntity"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/TodoEntity"
                }
              },
              "application/x-yaml": {
                "schema": {
                  "$ref": "#/components/schemas/TodoEntity"
                }
              },
              "text/yaml": {
                "schema": {
                  "$ref": "#/components/schemas/TodoEntity"
                }
              }
            }
          }
        }
      }
    },
    "/api/Todo/{id}": {
      "get": {
        "tags": [
          "Todo"
        ],
        "operationId": "Todo2",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/TodoEntity"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TodoEntity"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/TodoEntity"
                }
              },
              "application/x-yaml": {
                "schema": {
                  "$ref": "#/components/schemas/TodoEntity"
                }
              },
              "text/yaml": {
                "schema": {
                  "$ref": "#/components/schemas/TodoEntity"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Todo"
        ],
        "operationId": "Todo3",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TodoEntity"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/TodoEntity"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/TodoEntity"
              }
            },
            "application/x-yaml": {
              "schema": {
                "$ref": "#/components/schemas/TodoEntity"
              }
            },
            "text/yaml": {
              "schema": {
                "$ref": "#/components/schemas/TodoEntity"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Todo"
        ],
        "operationId": "Todo4",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/TodoEntity"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TodoEntity"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/TodoEntity"
                }
              },
              "application/x-yaml": {
                "schema": {
                  "$ref": "#/components/schemas/TodoEntity"
                }
              },
              "text/yaml": {
                "schema": {
                  "$ref": "#/components/schemas/TodoEntity"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "TodoEntity": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "title"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string",
            "maxLength": 200
          },
          "start": {
            "type": "string",
            "format": "date-time"
          },
          "end": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "finished": {
            "type": "boolean"
          }
        }
      }
    }
  }
}